example: refactor research
